Add files via upload
Wrote the Verilog code of the complete interleaver. It is an interleaver which is a fundamental
part of the turbo code design and plays critical role in the performance of turbo decoding.It consists of an
address generation unit, a ram, multiplexers, and a counter which works as an address and a reset unit. Here
AGU generates 12283 random addresses. Here for the first 12283 clock cycles the data is stored sequentially
in ram and after the datas are taken out randomly.